+++ src/app/app-store.module.ts
@@ -6,13 +6,13 @@ import { routerReducer, StoreRouterConnectingModule } from '@ngrx/router-store';
 import { HospitalCatReducer } from './HospitalCategorie/Store/reducer';
 import { DoctorCatReducer } from './doctorCategorie/Store/reducer';
 import { AppointementReducer } from './appointements/store/appointement.reducer';
-import { HospitalReducer } from './hospital/store/Reducer';
+import { HospitalReducer } from './hospitals/store/Reducer';
 import { doctorReducer } from './doctors/doctor-store/doctor.reducer';
 import { DoctorsEffect } from './doctors/doctor-store/doctor.effect';
 import { HospitalCatEffect } from './HospitalCategorie/Store/Effect';
 import { DoctorCatEffect } from './doctorCategorie/Store/Effect';
 import { AppointementEffect } from './appointements/store/appointement.effect';
-import { HospitalEffect } from './hospital/store/Effect';
+import { HospitalEffect } from './hospitals/store/Effect';
 import { OperationReducer } from './operations/store/operations.reducer';
 import { OperationCategoryReducer } from './operations/store/category.reducer';
 import { OpEffect } from './operations/store/operation.effect';
