+++ src/app/hospitals/hopital/hopital.component.html
