+++ src/app/hospitals/hospital.model.ts
