+++ src/app/hospitals/store/app-state.ts
